awq/tests/test_quantization.py
import unittest
from unittest.mock import patch, MagicMock
import os
from app.quantization import run_quantization, quantize_model
from app.config import Config

class TestQuantization(unittest.TestCase):
    def setUp(self):
        self.model_path = "/path/to/model"
        self.quant_config = {
            "zero_point": True,
            "q_group_size": 128,
            "w_bit": 4,
            "version": "GEMM"
        }

    @patch('app.quantization.load_safetensors_model')
    @patch('app.quantization.quantize_model')
    @patch('app.quantization.save_safetensors_model')
    def test_run_quantization_success(self, mock_save, mock_quantize, mock_load):
        # Setup
        mock_load.return_value = {"weights": [1, 2, 3]}
        mock_quantize.return_value = {"quantized_weights": [0.5, 1, 1.5]}

        # Run the function
        run_quantization(self.model_path, self.quant_config)

        # Assertions
        mock_load.assert_called_once_with(self.model_path)
        mock_quantize.assert_called_once_with({"weights": [1, 2, 3]}, self.quant_config)
        mock_save.assert_called_once()
        self.assertTrue("quantized_model.safetensors" in mock_save.call_args[0][1])

    @patch('app.quantization.load_safetensors_model')
    def test_run_quantization_failure(self, mock_load):
        # Setup
        mock_load.side_effect = Exception("Failed to load model")

        # Run the function and check for exception
        with self.assertRaises(Exception):
            run_quantization(self.model_path, self.quant_config)

    def test_quantize_model(self):
        # Setup
        model = {"weights": [1, 2, 3]}
        config = {"some_config": "value"}

        # Run the function
        result = quantize_model(model, config)

        # Assertions
        self.assertEqual(result, model, "Placeholder quantization should return the original model")

if __name__ == '__main__':
    unittest.main()
